Feat : UI StatsWindow Logic Add
보상?상점? 스탯창(잼 이용해서) 스탯 업그레이드 가능

+++ RogueLike/Assets/02.Scripts/Managers/DataManager.cs
@@ -76,7 +76,7 @@ public class DataManager : MonoBehaviour
         PlayerCurrentStats.attackSO.range += range;
     }
 
-    public void UpdatePlayerStatsData(int maxHealth ,int health, int speed)
+    public void UpdatePlayerStatsData(int maxHealth ,int health, float speed)
     {
         PlayerCurrentStats.maxHealth += maxHealth;
         PlayerCurrentStats.currentHealth += health;
